Effect required changes
