+++ src/index.html
@@ -3,32 +3,25 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-giJF6kkoqNQ00vy+HMDP7azOuL0xtbfIcaT9wjKHr8RbDVddVHyTfAAsrekwKmP1" crossorigin="anonymous">
-<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.3.0/font/bootstrap-icons.css">
-<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.3/css/all.min.css" integrity="sha512-iBBXm8fW90+nuLcSKlbmrPcLa0OT92xO1BIsZ+ywDWZCvqsWgccV3gFoRBv0z+8dLJgyAHIhR35VZc2oM/gI1w==" crossorigin="anonymous" referrerpolicy="no-referrer" />
-<script src="https://kit.fontawesome.com/a076d05399.js"></script>
-    <title>Todo List</title>
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.3/css/all.min.css" integrity="sha512-iBBXm8fW90+nuLcSKlbmrPcLa0OT92xO1BIsZ+ywDWZCvqsWgccV3gFoRBv0z+8dLJgyAHIhR35VZc2oM/gI1w==" crossorigin="anonymous" referrerpolicy="no-referrer" />
+    <title>TodoList</title>
 </head>
 <body>
-    
-    
-                <div class="wrapper">
-                    <div class="mini-section title">   
-                        <p>Today's To Do</p>
-                        <i class="fas fa-sync-alt"></i>
-                    </div>
-                    <div class="mini-section do">
-                        <input type="text" placeholder="Add to your list..." class="task">
-                        <i class="fas fa-level-down-alt fa-rotate-90"></i>
-                    </div>
-                    <div class="todoList">
-        
-                    </div>
-                    <div class="clear-all">
-                        <p>Clear all completed</p>
-                    </div>
-                </div>
-        
-<script src="main.js"></script>
+    <div class="wrapper">
+        <div class="mini-section title">   
+            <p>Today's To Do</p>
+            <i class="fas fa-sync-alt"></i>
+        </div>
+        <div class="mini-section do">
+            <input type="text" placeholder="Add to your list..." class="task">
+            <i class="fas fa-level-down-alt fa-rotate-90"></i>
+        </div>
+        <div class="todoList">
+            
+        </div>
+        <div class="clear-all">
+            <p>Clear all completed</p>
+        </div>
+    </div>
 </body>
 </html>
