+++ src/index.js
@@ -1,16 +1,17 @@
 /* eslint-disable import/prefer-default-export */
 /* eslint-disable import/no-cycle */
 /* eslint-disable import/no-duplicates */
-import './style.css';
+
 import {
   reloadContainer,
   addListeners,
   dragOver,
   list,
+  deleteItem,
   deleteAll,
   inputListener,
 } from './drag.js';
-import { deleteItem } from './addRemove.js';
+import './style.css';
 import { getContainer } from './complete.js';
 import editListener from './edit.js';
 
@@ -18,8 +19,7 @@ export const todoList = document.querySelector('.todoList');
 export const clearAll = document.querySelector('.clear-all');
 
 const task = document.querySelector('.task');
-export const displayList = (list) => {
-  todoList.innerHTML = '';
+export const showTask = (list) => {
   list.forEach((e) => {
     const lists = document.createElement('div');
     lists.classList.add('mini-section');
@@ -47,11 +47,10 @@ export const displayList = (list) => {
   });
 };
 
-displayList(list);
+showTask(list);
 getContainer(todoList);
 export const items = document.querySelectorAll('.item');
 export const del = document.querySelector('.fa-trash');
-
 inputListener(task);
 editListener();
 addListeners(items);
